Normalize impl names
