+++ impls/rust/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "ip-rust"
+name = "ipcheck"
 version = "0.0.0"
 authors = ["Ashley Mannix <kodraus@hey.com>", "Christiaan Dirkx <christiaan@dirkx.email>"]
 publish = false
 edition = "2018"
 
 [[bin]]
-name = "ip-rust"
+name = "ipcheck"
 path = "main.rs"
 
 [dependencies.serde_json]
