+++ impls/rust/main.rs
